requirements.txt
BeautifulSoup==3.2.1
Django==1.8.11
Pillow==2.7.0
Unidecode==0.04.18
argparse==1.3.0
beautifulsoup4==4.3.2
boto==2.36.0
csscompressor==0.9.3
days==0.0.3
distribute==0.6.38
dj-database-url==0.1.2
django-ace==1.0.2
django-ace-overlay==0.3
django-appconf==1.0.1
django-batch-uploader==0.12.1
-e git+https://github.com/ninapavlich/django-carbon.git@c43d535f01d4c3a9f8a3b591859a25fc5813c490#egg=django_carbon-dev
django-ckeditor==5.0.2
django-ckeditor-improved==1.2.3
django-debug-toolbar==1.4
django-extensions==1.5.0
django-extra-views==0.6.5
django-grappelli==2.7.1
django-heroku-memcacheify==0.8
django-heroku-postgresify==0.3
django-herokuify==1.0.pre3
django-htmlmin==0.7.0
django-imagekit==3.2.7
#-e git+https://github.com/ninapavlich/django-imagekit-cropper.git@1.14#egg=imagekit_cropper-dev
django-imagekit-cropper==1.14
django-inline-wrestler==4.4
django-localflavor==1.1
django-object-actions==0.5.1
django-pylibmc==0.5.0
django-pylibmc-sasl==0.2.4
django-reversion==1.8.7
django-robots==1.1
django-share-me-share-me==1.0
django-storages==1.1.8
django-unsaved-changes==0.9
feedparser==5.2.1
gunicorn==19.2.1
heroku==0.1.4
html5lib==1.0b3
httplib2==0.9.2
libsass==0.7.0
lxml==3.5.0
newrelic==2.44.0.36
oauthlib==1.0.3
parsedate==0.0.1
pilkit==1.1.13
ply==3.4
psycopg2==2.6.1
pylibmc==1.4.1
pyparsing==2.0.3
python-dateutil==1.5
python-instagram==1.3.2
pytz==2015.7
raven==5.8.1
requests==2.8.1
requests-oauthlib==0.5.0
-e git+https://github.com/ninapavlich/simple-social-suture.git@0.4#egg=simple_social_suture-dev
simplejson==3.8.2
six==1.10.0
slimit==0.8.1
sqlparse==0.1.16
twitter-text-python==1.1.0
twython==3.3.0
wsgiref==0.1.2
